# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for MyCar application",
        "title": "MyCar API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Login user and get authentication token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns JWT token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
...
        "/api/moto": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new moto for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "motos"
                ],
                "summary": "Create a new moto",
                "parameters": [
                    {
                        "description": "Moto creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MotoRq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns ID of created moto",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
        "/api/moto/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает мотоцикл по идентификатору для аутентифицированного пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "motos"
                ],
                "summary": "Получить мотоцикл по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moto ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные мотоцикла",
                        "schema": {
                            "$ref": "#/definitions/moto.Moto"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "404": {
                        "description": "Мотоцикл не найден",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing moto for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "motos"
                ],
                "summary": "Update a moto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moto ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Moto update data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MotoRq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moto updated successfully"
                    },
                    "400": {
                        "description": "Invalid request data or ID",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "404": {
                        "description": "Moto not found",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an existing moto for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "motos"
                ],
                "summary": "Delete a moto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moto ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moto deleted successfully"
                    },
                    "400": {
                        "description": "Invalid moto ID",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "404": {
                        "description": "Moto not found",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AuthReq": {
            "description": "User authentication credentials",
            "type": "object",
            "required": [
                "Login",
                "Password"
            ],
            "properties": {
                "Login": {
                    "type": "string",
                    "example": "user123"
                },
                "Password": {
                    "type": "string",
                    "example": "securePassword123"
                }
            }
        },
...
        "handler.MotoRq": {
            "type": "object",
            "required": [
                "brand",
                "category",
                "fuel_type",
                "plate",
                "transmission_type",
                "vin",
                "year"
            ],
            "properties": {
                "brand": {
                    "$ref": "#/definitions/moto.Brand"
                },
                "category": {
                    "$ref": "#/definitions/moto.CategoryKind"
                },
                "fuel_type": {
                    "$ref": "#/definitions/vehicle.FuelType"
                },
                "plate": {
                    "type": "string"
                },
                "transmission_type": {
                    "$ref": "#/definitions/moto.TransmissionTypeKind"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "handler.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "model.AuditFields": {
            "type": "object",
            "properties": {
                "createdAtUtc": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "modifiedAtUtc": {
                    "type": "string"
                },
                "modifiedBy": {
                    "type": "string"
                }
            }
        },
        "moto.Brand": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29,
                30,
                31,
                32,
                33,
                34,
                35,
                36,
                37,
                38,
                39,
                40,
                41,
                42,
                43,
                44,
                45,
                46,
                47,
                48,
                49,
                50,
                51,
                52,
                53,
                54,
                55,
                56,
                57,
                58,
                59,
                60,
                61,
                62,
                63,
                64,
                65,
                66,
                67,
                68,
                69,
                70,
                71,
                72,
                73,
                74,
                75,
                76,
                77,
                78,
                79,
                80,
                81,
                82,
                83,
                84,
                85,
                86,
                87,
                88,
                89,
                90,
                91,
                92,
                93,
                94,
                95,
                96,
                97,
                98,
                99,
                100,
                101,
                102,
                103,
                104,
                105,
                106,
                107,
                108,
                109,
                110,
                111,
                112,
                113,
                114,
                115,
                116,
                117,
                118,
                119,
                120,
                121,
                122,
                123,
                124,
                125,
                126,
                127,
                128,
                129,
                130,
                131,
                132,
                133,
                134,
                135,
                136,
                137,
                138,
                139,
                140,
                141,
                142,
                143,
                144,
                145,
                146,
                147,
                148,
                149,
                150,
                151,
                152,
                153,
                154,
                155,
                156,
                157,
                158,
                159,
                160,
                161,
                162,
                163,
                164,
                165,
                166,
                167,
                168,
                169,
                170,
                171,
                172,
                173,
                174,
                175,
                176,
                177,
                178,
                179,
                180,
                181,
                182,
                183,
                184,
                185,
                186,
                187,
                188,
                189,
                190,
                191,
                192,
                193,
                194,
                195,
                196,
                197,
                198,
                199,
                200,
                201,
                202,
                203,
                204,
                205,
                206,
                207,
                208,
                209,
                210,
                211,
                212,
                213,
                214,
                215,
                216,
                217,
                218,
                219,
                220,
                221,
                222,
                223,
                224,
                225,
                226,
                227,
                228
            ],
            "x-enum-varnames": [
                "UnknownBrand",
                "Aermacchi",
                "AJP",
                "Alfer",
                "Amazonas",
                "AmericanEagle",
                "AmericanIronHorse",
                "Anaig",
                "Apollo",
                "Aprilia",
                "Ardie",
                "Armada",
                "AstonMartin",
                "Ataki",
                "ATK",
                "Avantis",
                "Bajaj",
                "Baltmotors",
                "Bamx",
                "Bars",
                "Bashan",
                "Benda",
                "Benelli",
                "Beta",
                "Bianchi",
                "Bifei",
                "BigBearChoppers",
                "BigDogMotorcycles",
                "Blata",
                "BoomTrikes",
                "BossHoss",
                "Brabus",
                "BRP",
                "BRZ",
                "BSA",
                "BSE",
                "Buell",
                "Cagiva",
                "Carver",
                "CCM",
                "Centurion",
                "CFMoto",
                "CH_Racing",
                "ChangJiang",
                "Cobra",
                "CPI",
                "CR_S",
                "Cronus",
                "CRZ",
                "Curtiss",
                "Cyclone",
                "CZ",
                "Darex",
                "Dayun",
                "Deller",
                "Derbi",
                "DesertRaven",
                "Draxter",
                "Energica",
                "Eurotex",
                "EVA",
                "ExpertMoto",
                "Falcon",
                "FidelisEtFortis",
                "Fighter",
                "Forsage",
                "Fosti",
                "Fuego",
                "FullCrew",
                "Futong",
                "FXmoto",
                "Gmax",
                "Gaokin",
                "GasGas",
                "Geon",
                "Gilera",
                "GR",
                "Groza",
                "Gryphon",
                "GXmoto",
                "Hasky",
                "Hero",
                "Highland",
                "Horwin",
                "Huangpeng",
                "Husaberg",
                "Husqvarna",
                "Hyosung",
                "Indian",
                "Irbis",
                "Iride",
                "ItalianVolt",
                "Jawa",
                "JHL",
                "JiangsuSunhou",
                "JMC",
                "JohnnyPag",
                "Jose",
                "K2R",
                "Keeway",
                "Kews",
                "KinlonMotors",
                "Koshine",
                "Kove",
                "KXD",
                "Kymco",
                "Lem",
                "Lifan",
                "Lito",
                "Loncin",
                "Magni",
                "Mikilon",
                "Minsk",
                "Montesa",
                "Motax",
                "MotoGuzzi",
                "MotoMorini",
                "Mowgli",
                "MVAgusta",
                "MZ",
                "Nexus",
                "Nicot",
                "Nitro",
                "NIU",
                "Norton",
                "Ohvale",
                "Omaks",
                "OrangeCountyChoppers",
                "Orion",
                "Oxo",
                "Pannonia",
                "Patron",
                "Pitrace",
                "PitsterPro",
                "Polaris",
                "Polini",
                "Procida",
                "Progasi",
                "Promax",
                "PWRRacing",
                "QJMotor",
                "Racer",
                "RegalRaptor",
                "Regulmoto",
                "Rewaco",
                "Rhino",
                "Rieju",
                "Rmoto",
                "Rockot",
                "Roliz",
                "RoyalEnfield",
                "S2Motors",
                "Sachs",
                "Sagitta",
                "Saxon",
                "Scanmoto",
                "Scorpa",
                "Sharmax",
                "Sherco",
                "Shineray",
                "SimarglElektro",
                "Simson",
                "Skygo",
                "Sonik",
                "SPR",
                "SSSR",
                "Stark",
                "Stels",
                "Stingray",
                "SuperSoco",
                "SurRon",
                "Swift",
                "SWM",
                "Sym",
                "Tacita",
                "Talaria",
                "TaoMotor",
                "Tinbot",
                "Titan",
                "TMRacing",
                "TMBK",
                "Triumph",
                "TRRS",
                "TVS",
                "UM",
                "Upbeat",
                "Vanderhall",
                "Vertigo",
                "Victory",
                "Vinto",
                "Viper",
                "VMC",
                "Voge",
                "Voxan",
                "Wanderer",
                "Wanqiang",
                "Wels",
                "WhiteSiberia",
                "Wuxi",
                "XGZ",
                "Xingyue",
                "Xmotos",
                "Yacota",
                "Yamasaki",
                "Yangtze",
                "YCF",
                "Yinxiang",
                "Zero",
                "ZIPMotors",
                "ZM",
                "Zongshen",
                "Zontes",
                "Zundapp",
                "Zuum",
                "Zuumav",
                "Gyurza",
                "Dnepr",
                "ZID",
                "IZH",
                "IzhTechMash",
                "Kovrovets",
                "LvivMotozavod",
                "MMZ",
                "Motoalfa",
                "Motomir",
                "PMZ",
                "RMZ",
                "TMZ",
                "Ural"
            ]
        },
        "moto.Category": {
            "type": "object"
        },
        "moto.CategoryKind": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29
            ],
            "x-enum-varnames": [
                "CategoryKindNone",
                "Offroad",
                "Road",
                "Cruisers",
                "Sports",
                "Supermoto",
                "Trikes",
                "Tourism",
                "Allround",
                "Enduro",
                "DualPurpose",
                "SportEnduro",
                "TouringEnduro",
                "Naked",
                "RoadBike",
                "Classic",
                "Custom",
                "Cruiser",
                "Chopper",
                "Cross",
                "Speedway",
                "Kids",
                "Minibike",
                "Pitbike",
                "Trial",
                "Sportbike",
                "SportTouring",
                "Supersport",
                "Trike",
                "ThreeWheeler"
            ]
        },
        "moto.Moto": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/model.AuditFields"
                },
                "brand": {
                    "$ref": "#/definitions/moto.Brand"
                },
                "category": {
                    "$ref": "#/definitions/moto.Category"
                },
                "fuelType": {
                    "$ref": "#/definitions/vehicle.FuelType"
                },
                "id": {
                    "type": "string"
                },
                "plate": {
                    "type": "string"
                },
                "transmissionType": {
                    "$ref": "#/definitions/moto.TransmissionType"
                },
                "userId": {
                    "type": "string"
                },
                "vehicleType": {
                    "$ref": "#/definitions/vehicle.Type"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "moto.TransmissionType": {
            "type": "object",
            "properties": {
                "id": {
                    "$ref": "#/definitions/moto.TransmissionTypeKind"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "$ref": "#/definitions/moto.TransmissionTypeKind"
                }
            }
        },
        "moto.TransmissionTypeKind": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-varnames": [
                "TransmissionTypeUnknown",
                "TransmissionTypeAuto",
                "TransmissionTypeAutomatic",
                "TransmissionTypeRobot",
                "TransmissionTypeCVT",
                "TransmissionTypeManual"
            ]
        },
        "vehicle.FuelType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7
            ],
            "x-enum-varnames": [
                "UnknownFuelType",
                "Petrol",
                "PetrolPropaneButane",
                "PetrolMethane",
                "PetrolHybrid",
                "Diesel",
                "DieselHybrid",
                "Electric"
            ]
        },
        "vehicle.Type": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "UnknownVehicleType",
                "Car",
                "Motorcycle"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}